refactor: '인기 테마 조회'기능의 비즈니스 로직을 서비스가 담당하도록 책임 재분배

+++ src/main/java/roomescape/repository/dto/ReservedTheme.java
@@ -7,6 +7,6 @@ public record ReservedTheme(
         String name,
         String description,
         String imageUrl,
-        int reservationCount
+        long reservationCount
 ) {
 }
